+++ docs/docs/resources/changelog.md
@@ -18,12 +18,14 @@ During the alpha period, things might break! We take breaking changes very serio
   - `wa-input` => `input`
   - `wa-change` => `change`
   - `wa-blur` => `blur` (this event will no longer bubble, use `focusout` for a bubbling version)
-  - `wa-focus` => `focus` (this event will no longer bubble)
+  - `wa-focus` => `focus` (this event will no longer bubble, use `focusin` for a bubbling version)
 - Added `.wa-callout` utility class
 - Added the `orientation` attribute to `<wa-radio-group>` to support vertical and horizontal radio items
 - Added docs for visual tests
 - Added docs on how to cherry-pick native styles
-- Changed `variant` behavior so that nested components with `variant` support inherit the `variant` set on their ancestor
+- Changed the behavior of the `variant` and `size` attributes so that nested components that support these attributes but do not have them set inherit the values set on their ancestors. Additionally:
+  - Added `size` attribute to `<wa-dropdown>`, `<wa-button-group>`, `<wa-menu>`, `<wa-rating>`, `<wa-card>`
+  - Added `variant` attribute to `<wa-button-group>`
 - Fixed a bug in `<wa-tab-group>` that prevented nested tab groups from working properly
 - Fixed slot names for `show-password-icon` and `hide-password-icon` in `<wa-input>` to more intuitively represent their functions
 - Fixed a bug in `<wa-textarea>` that caused empty controls to submit a value if the initial value was deleted a certain way